IPC FW3.0 | build project | sys_server; Description:add check of schedule ; Risk: low; Affected Module: no for now; Test: compiled on Win7 and Ubuntu 10.04; Code Review: no;
git-svn-id: https://192.168.0.21/svn/ipc_fw3.x_core/branches/20131125@1344 dd879d07-adaa-4757-b7d0-008d00b626b4

+++ application/sys_env_types.h
@@ -1562,9 +1562,9 @@ typedef struct tagLinkAlarmExtInfo
 	uint8_t  s_IoNum;//bit: 0-alarm port 0, 1-alarm port 1, 2-alarm port 2, 3-alarm port 3
 	uint8_t  s_OperateCmd;//when link ptz:0--none, 1--preset ,2--cruise, 3--scan
 	uint16_t s_OperateSeqNum;//when link ptz: seqno: 0(home),1,2,3...
-	uint8_t  s_PtzDelayTime;//unit:s  ptz delay time
-	uint8_t  s_DelayTime;//unit:s white light delay time
-	uint8_t  s_Reserved[6];	
+	uint16_t  s_PtzDelayTime;//unit:s  ptz delay time
+	uint16_t  s_DelayTime;//unit:s white light delay time
+	uint16_t  s_Reserved[6];	
 }SysPkgLinkAlarmExtInfo;
